Log response body when it can't be converted to JSON
When the response was somehting like 'Unauthorized' or anything other
than a JSON, we'd get a Json exception because that string couldn't be
converted to a JSON object.

On server, the API returns an HTML instead of a string, and the whole
HTML content will be logged to the errors.

This approach prevents users from seeing weird errors in their logs such
as 'org.json.JSONException: A JSONArray text must start with '[' at 1
[character 2 line 1]' and gives them a proper error message, so they can
either fix it by themselves or find help with support.